Add ES6 module
