+++ index.html
@@ -62,7 +62,8 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
     crossorigin="anonymous"></script>
-  <script src="js/script.js"></script>
+  <script type="module" src="js/book.js"></script>
+  <script type="module" src="js/script.js"></script>
 </body>
 
 </html>
